requirements.txt
Flask
flask-cors
python-dotenv
youtube-transcript-api
groq
yt-dlp
requests